tercer commit: rest de player, funcional

+++ RetoCrudBackend/src/main/java/co/com/adrianafranklin/RetoCrudBackend/Repository/RepositoryPlayer.java
@@ -3,5 +3,5 @@ package co.com.adrianafranklin.RetoCrudBackend.Repository;
 import co.com.adrianafranklin.RetoCrudBackend.Entitys.Player;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RepositoryPlayer extends CrudRepository<Player, Long> {
+public interface RepositoryPlayer extends CrudRepository<Player, Integer> {
 }
